+++ typings/custom-types.d.ts
@@ -1,5 +1,4 @@
 declare module '@google-cloud/connect-datastore';
-declare module '@google-cloud/tasks';
 declare module 'helmet-csp';
 declare module 'passport-google-oauth20';
 declare module 'passport-openidconnect';